Rename CommonState -> State and Framework -> Rules
