+++ packages/fmg-rock-paper-scissors/contracts/RockPaperScissorsState.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.4.18;
 
-import "fmg-core/contracts/CommonState.sol";
+import "fmg-core/contracts/State.sol";
 
 library RockPaperScissorsState {
     enum PositionType { Start, RoundProposed, RoundAccepted, Reveal, Concluded }
@@ -17,42 +17,42 @@ library RockPaperScissorsState {
     // [160 - 191] bytes32 salt
 
     function positionType(bytes _state) public pure returns (PositionType _positionType) {
-        uint offset = CommonState.gameStateOffset(_state);
+        uint offset = State.gameStateOffset(_state);
         assembly {
             _positionType := mload(add(_state, offset))
         }
     }
 
     function stake(bytes _state) public pure returns (uint256 _stake) {
-        uint offset = CommonState.gameStateOffset(_state) + 32;
+        uint offset = State.gameStateOffset(_state) + 32;
         assembly {
             _stake := mload(add(_state, offset))
         }
     }
 
     function preCommit(bytes _state) public pure returns (bytes32 _preCommit) {
-        uint offset = CommonState.gameStateOffset(_state) + 64;
+        uint offset = State.gameStateOffset(_state) + 64;
         assembly {
             _preCommit := mload(add(_state, offset))
         }
     }
 
     function bPlay(bytes _state) public pure returns (Play _aPlay) {
-        uint offset = CommonState.gameStateOffset(_state) + 96;
+        uint offset = State.gameStateOffset(_state) + 96;
         assembly {
             _aPlay := mload(add(_state, offset))
         }
     }
 
     function aPlay(bytes _state) public pure returns (Play _bPlay) {
-        uint offset = CommonState.gameStateOffset(_state) + 128;
+        uint offset = State.gameStateOffset(_state) + 128;
         assembly {
             _bPlay := mload(add(_state, offset))
         }
     }
 
     function salt(bytes _state) public pure returns (bytes32 _salt) {
-        uint offset = CommonState.gameStateOffset(_state) + 160;
+        uint offset = State.gameStateOffset(_state) + 160;
         assembly {
             _salt := mload(add(_state, offset))
         }
@@ -60,10 +60,10 @@ library RockPaperScissorsState {
 
     // utility functions
     function aResolution(bytes _state) public pure returns (uint256) {
-        return CommonState.resolution(_state)[0];
+        return State.resolution(_state)[0];
     }
 
     function bResolution(bytes _state) public pure returns (uint256 _bBal) {
-        return CommonState.resolution(_state)[1];
+        return State.resolution(_state)[1];
     }
 }
